#1531 - extract out users & related dois components & write unit tests for related DOIs
