+++ packages/datagateway-download/src/DOIGenerationForm/DOIGenerationForm.component.test.tsx
@@ -15,6 +15,9 @@ import { DownloadSettingsContext } from '../ConfigProvider';
 import { mockCartItems, mockedSettings } from '../testData';
 import {
   checkUser,
+  DOIRelationType,
+  DOIResourceType,
+  fetchDOI,
   getCartUsers,
   isCartMintable,
   mintCart,
@@ -34,9 +37,6 @@ jest.mock('datagateway-common', () => {
     __esModule: true,
     ...originalModule,
     fetchDownloadCart: jest.fn(),
-    readSciGatewayToken: jest.fn(() => ({
-      username: '1',
-    })),
   };
 });
 
@@ -49,6 +49,7 @@ jest.mock('../downloadApi', () => {
     getCartUsers: jest.fn(),
     checkUser: jest.fn(),
     mintCart: jest.fn(),
+    fetchDOI: jest.fn(),
   };
 });
 
@@ -107,21 +108,24 @@ describe('DOI generation form component', () => {
         email: 'user1@example.com',
         affiliation: 'Example Uni',
       },
-      {
-        id: 2,
-        name: '2',
-        fullName: 'User 2',
-        email: 'user2@example.com',
-        affiliation: 'Example 2 Uni',
-      },
     ]);
 
     (checkUser as jest.MockedFunction<typeof checkUser>).mockResolvedValue({
-      id: 3,
-      name: '3',
-      fullName: 'User 3',
-      email: 'user3@example.com',
-      affiliation: 'Example 3 Uni',
+      id: 2,
+      name: '2',
+      fullName: 'User 2',
+      email: 'user2@example.com',
+      affiliation: 'Example 2 Uni',
+    });
+
+    (fetchDOI as jest.MockedFunction<typeof fetchDOI>).mockResolvedValue({
+      id: '1',
+      type: 'DOI',
+      attributes: {
+        doi: 'related.doi.1',
+        titles: [{ title: 'Related DOI 1' }],
+        url: 'www.example.com',
+      },
     });
   });
 
@@ -187,7 +191,7 @@ describe('DOI generation form component', () => {
     );
   });
 
-  it('should let the user delete users (but not delete the logged in user)', async () => {
+  it('should not let the user submit a mint request if required fields are missing but can submit once all are filled in', async () => {
     renderComponent();
 
     // accept data policy
@@ -195,168 +199,41 @@ describe('DOI generation form component', () => {
       screen.getByRole('button', { name: 'acceptDataPolicy.accept' })
     );
 
+    // missing title
     expect(
-      within(screen.getByRole('table', { name: 'DOIGenerationForm.creators' }))
-        .getAllByRole('row')
-        .slice(1) // ignores the header row
-    ).toHaveLength(2);
-    expect(
-      screen.getByRole('cell', { name: 'user2@example.com' })
-    ).toBeInTheDocument();
+      screen.getByRole('button', { name: 'DOIGenerationForm.generate_DOI' })
+    ).toBeDisabled();
 
-    const userDeleteButtons = screen.getAllByRole('button', {
-      name: 'DOIGenerationForm.delete_creator',
-    });
-    expect(userDeleteButtons[0]).toBeDisabled();
-
-    await user.click(userDeleteButtons[1]);
-
-    expect(
-      within(screen.getByRole('table', { name: 'DOIGenerationForm.creators' }))
-        .getAllByRole('row')
-        .slice(1)
-    ).toHaveLength(1);
-    expect(
-      screen.getByRole('cell', { name: 'Example Uni' })
-    ).toBeInTheDocument();
-  });
-
-  it('should let the user add creators (but not duplicate users or if checkUser fails)', async () => {
-    renderComponent();
-
-    // accept data policy
-    await user.click(
-      screen.getByRole('button', { name: 'acceptDataPolicy.accept' })
+    await user.type(
+      screen.getByRole('textbox', { name: 'DOIGenerationForm.title' }),
+      't'
     );
 
+    // missing description
     expect(
-      within(screen.getByRole('table', { name: 'DOIGenerationForm.creators' }))
-        .getAllByRole('row')
-        .slice(1) // ignores the header row
-    ).toHaveLength(2);
+      screen.getByRole('button', { name: 'DOIGenerationForm.generate_DOI' })
+    ).toBeDisabled();
+
+    await user.type(
+      screen.getByRole('textbox', { name: 'DOIGenerationForm.description' }),
+      'd'
+    );
+
+    // missing cart users
 
     await user.type(
       screen.getByRole('textbox', { name: 'DOIGenerationForm.username' }),
-      '3'
-    );
-
-    await user.click(
-      screen.getByRole('button', { name: 'DOIGenerationForm.add_creator' })
-    );
-
-    expect(
-      within(screen.getByRole('table', { name: 'DOIGenerationForm.creators' }))
-        .getAllByRole('row')
-        .slice(1) // ignores the header row
-    ).toHaveLength(3);
-    expect(screen.getByRole('cell', { name: 'User 3' })).toBeInTheDocument();
-    expect(screen.getAllByRole('cell', { name: 'Creator' }).length).toBe(3);
-
-    // test errors on duplicate user
-    await user.type(
-      screen.getByRole('textbox', { name: 'DOIGenerationForm.username' }),
-      '3'
-    );
-
-    await user.click(
-      screen.getByRole('button', { name: 'DOIGenerationForm.add_creator' })
-    );
-
-    expect(
-      within(screen.getByRole('table', { name: 'DOIGenerationForm.creators' }))
-        .getAllByRole('row')
-        .slice(1) // ignores the header row
-    ).toHaveLength(3);
-    expect(screen.getByText('Cannot add duplicate user')).toBeInTheDocument();
-    expect(
-      screen.getByRole('textbox', { name: 'DOIGenerationForm.username' })
-    ).toHaveValue('');
-
-    // test errors with various API error responses
-    (checkUser as jest.MockedFunction<typeof checkUser>).mockRejectedValueOnce({
-      response: { data: { detail: 'error msg' }, status: 404 },
-    });
-
-    await user.type(
-      screen.getByRole('textbox', { name: 'DOIGenerationForm.username' }),
-      '4'
-    );
-
-    await user.click(
-      screen.getByRole('button', { name: 'DOIGenerationForm.add_creator' })
-    );
-
-    expect(await screen.findByText('error msg')).toBeInTheDocument();
-    expect(
-      within(screen.getByRole('table', { name: 'DOIGenerationForm.creators' }))
-        .getAllByRole('row')
-        .slice(1) // ignores the header row
-    ).toHaveLength(3);
-
-    (checkUser as jest.MockedFunction<typeof checkUser>).mockRejectedValue({
-      response: { data: { detail: [{ msg: 'error msg 2' }] }, status: 404 },
-    });
-    await user.click(
-      screen.getByRole('button', { name: 'DOIGenerationForm.add_creator' })
-    );
-
-    expect(await screen.findByText('error msg 2')).toBeInTheDocument();
-    expect(
-      within(screen.getByRole('table', { name: 'DOIGenerationForm.creators' }))
-        .getAllByRole('row')
-        .slice(1) // ignores the header row
-    ).toHaveLength(3);
-
-    (checkUser as jest.MockedFunction<typeof checkUser>).mockRejectedValueOnce({
-      response: { status: 422 },
-    });
-    await user.click(
-      screen.getByRole('button', { name: 'DOIGenerationForm.add_creator' })
-    );
-
-    expect(await screen.findByText('Error')).toBeInTheDocument();
-    expect(
-      within(screen.getByRole('table', { name: 'DOIGenerationForm.creators' }))
-        .getAllByRole('row')
-        .slice(1) // ignores the header row
-    ).toHaveLength(3);
-  });
-
-  it('should let the user add contributors & select their contributor type', async () => {
-    renderComponent();
-
-    // accept data policy
-    await user.click(
-      screen.getByRole('button', { name: 'acceptDataPolicy.accept' })
-    );
-
-    expect(
-      within(screen.getByRole('table', { name: 'DOIGenerationForm.creators' }))
-        .getAllByRole('row')
-        .slice(1) // ignores the header row
-    ).toHaveLength(2);
-
-    await user.type(
-      screen.getByRole('textbox', { name: 'DOIGenerationForm.username' }),
-      '3'
+      '2'
     );
 
     await user.click(
       screen.getByRole('button', { name: 'DOIGenerationForm.add_contributor' })
     );
 
+    // missing contributor type
     expect(
-      within(screen.getByRole('table', { name: 'DOIGenerationForm.creators' }))
-        .getAllByRole('row')
-        .slice(1) // ignores the header row
-    ).toHaveLength(3);
-    expect(screen.getByRole('cell', { name: 'User 3' })).toBeInTheDocument();
-
-    expect(
-      screen.getByRole('button', {
-        name: /DOIGenerationForm.creator_type/i,
-      })
-    ).toBeInTheDocument();
+      screen.getByRole('button', { name: 'DOIGenerationForm.generate_DOI' })
+    ).toBeDisabled();
 
     await user.click(
       screen.getByRole('button', {
@@ -367,20 +244,38 @@ describe('DOI generation form component', () => {
       await screen.findByRole('option', { name: 'DataCollector' })
     );
 
-    expect(screen.queryByRole('option')).not.toBeInTheDocument();
-    // check that the option is actually selected in the table even after the menu closes
-    expect(screen.getByText('DataCollector')).toBeInTheDocument();
-
-    // check users and their contributor types get passed correctly to API
     await user.type(
-      screen.getByRole('textbox', { name: 'DOIGenerationForm.title' }),
-      't'
+      screen.getByRole('textbox', { name: 'DOIGenerationForm.related_doi' }),
+      '1'
     );
 
-    await user.type(
-      screen.getByRole('textbox', { name: 'DOIGenerationForm.description' }),
-      'd'
+    await user.click(
+      screen.getByRole('button', { name: 'DOIGenerationForm.add_related_doi' })
     );
+
+    // missing relationship type
+    expect(
+      screen.getByRole('button', { name: 'DOIGenerationForm.generate_DOI' })
+    ).toBeDisabled();
+
+    await user.click(
+      screen.getByRole('button', {
+        name: /DOIGenerationForm.related_doi_relationship/i,
+      })
+    );
+    await user.click(await screen.findByRole('option', { name: 'IsCitedBy' }));
+
+    // missing resource type
+    expect(
+      screen.getByRole('button', { name: 'DOIGenerationForm.generate_DOI' })
+    ).toBeDisabled();
+
+    await user.click(
+      screen.getByRole('button', {
+        name: /DOIGenerationForm.related_doi_resource_type/i,
+      })
+    );
+    await user.click(await screen.findByRole('option', { name: 'Journal' }));
 
     await user.click(
       screen.getByRole('button', { name: 'DOIGenerationForm.generate_DOI' })
@@ -392,12 +287,103 @@ describe('DOI generation form component', () => {
         title: 't',
         description: 'd',
         creators: [
-          { username: '2', contributor_type: 'Creator' },
-          { username: '3', contributor_type: 'DataCollector' },
+          { username: '1', contributor_type: 'Creator' },
+          { username: '2', contributor_type: 'DataCollector' },
+        ],
+        related_items: [
+          {
+            title: 'Related DOI 1',
+            fullReference: '',
+            relatedIdentifier: 'related.doi.1',
+            relatedIdentifierType: 'DOI',
+            relationType: DOIRelationType.IsCitedBy,
+            resourceType: DOIResourceType.Journal,
+          },
         ],
       },
       expect.any(Object)
     );
+  });
+
+  it('should not let the user submit a mint request if cart fails to load', async () => {
+    (
+      fetchDownloadCart as jest.MockedFunction<typeof fetchDownloadCart>
+    ).mockRejectedValue({ message: 'error' });
+    renderComponent();
+
+    // accept data policy
+    await user.click(
+      screen.getByRole('button', { name: 'acceptDataPolicy.accept' })
+    );
+
+    await user.type(
+      screen.getByRole('textbox', { name: 'DOIGenerationForm.title' }),
+      't'
+    );
+
+    await user.type(
+      screen.getByRole('textbox', { name: 'DOIGenerationForm.description' }),
+      'd'
+    );
+
+    // missing cart
+    expect(
+      screen.getByRole('button', { name: 'DOIGenerationForm.generate_DOI' })
+    ).toBeDisabled();
+  });
+
+  it('should not let the user submit a mint request if cart is empty', async () => {
+    (
+      fetchDownloadCart as jest.MockedFunction<typeof fetchDownloadCart>
+    ).mockResolvedValue([]);
+    renderComponent();
+
+    // accept data policy
+    await user.click(
+      screen.getByRole('button', { name: 'acceptDataPolicy.accept' })
+    );
+
+    await user.type(
+      screen.getByRole('textbox', { name: 'DOIGenerationForm.title' }),
+      't'
+    );
+
+    await user.type(
+      screen.getByRole('textbox', { name: 'DOIGenerationForm.description' }),
+      'd'
+    );
+
+    // empty cart
+    expect(
+      screen.getByRole('button', { name: 'DOIGenerationForm.generate_DOI' })
+    ).toBeDisabled();
+  });
+
+  it('should not let the user submit a mint request if no users selected', async () => {
+    (
+      getCartUsers as jest.MockedFunction<typeof getCartUsers>
+    ).mockResolvedValue([]);
+    renderComponent();
+
+    // accept data policy
+    await user.click(
+      screen.getByRole('button', { name: 'acceptDataPolicy.accept' })
+    );
+
+    await user.type(
+      screen.getByRole('textbox', { name: 'DOIGenerationForm.title' }),
+      't'
+    );
+
+    await user.type(
+      screen.getByRole('textbox', { name: 'DOIGenerationForm.description' }),
+      'd'
+    );
+
+    // no users
+    expect(
+      screen.getByRole('button', { name: 'DOIGenerationForm.generate_DOI' })
+    ).toBeDisabled();
   });
 
   it('should let the user change cart tabs', async () => {
